@@ -8,8 +8,8 @@ import lombok.Getter;
 import lombok.extern.jackson.Jacksonized;
 import no.fintlabs.validation.constraints.IntegrationAndMetadataMatches;
 
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 
 import java.time.Instant;
 
