@@ -1,6 +1,6 @@
 package no.fintlabs.validation.groups;
 
-import javax.validation.GroupSequence;
+import jakarta.validation.GroupSequence;
 
 @GroupSequence({ValueParsability.class, InstanceValueKeys.class, InstanceValueTypes.class})
 public interface Completed {
