@@ -1,11 +1,15 @@
 package no.fintlabs.kafka;
 
 import no.fintlabs.ConfigurationService;
-import no.fintlabs.kafka.common.topic.TopicCleanupPolicyParameters;
+import no.fintlabs.kafka.consuming.ErrorHandlerConfiguration;
+import no.fintlabs.kafka.consuming.ErrorHandlerFactory;
 import no.fintlabs.kafka.requestreply.ReplyProducerRecord;
-import no.fintlabs.kafka.requestreply.RequestConsumerFactoryService;
-import no.fintlabs.kafka.requestreply.topic.RequestTopicNameParameters;
+import no.fintlabs.kafka.requestreply.RequestListenerConfiguration;
+import no.fintlabs.kafka.requestreply.RequestListenerContainerFactory;
 import no.fintlabs.kafka.requestreply.topic.RequestTopicService;
+import no.fintlabs.kafka.requestreply.topic.configuration.RequestTopicConfiguration;
+import no.fintlabs.kafka.requestreply.topic.name.RequestTopicNameParameters;
+import no.fintlabs.kafka.topic.name.TopicNamePrefixParameters;
 import no.fintlabs.model.configuration.dtos.ConfigurationDto;
 import no.fintlabs.model.configuration.dtos.ObjectMappingDto;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
@@ -13,31 +17,56 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.kafka.listener.ConcurrentMessageListenerContainer;
 
+import java.time.Duration;
+
 @Configuration
 public class ConfigurationRequestConsumerConfiguration {
+
+    private static final Duration RETENTION_TIME = Duration.ofMinutes(5);
 
     @Bean
     ConcurrentMessageListenerContainer<String, Long>
     configurationByConfigurationIdRequestConsumer(
             RequestTopicService requestTopicService,
-            RequestConsumerFactoryService requestConsumerFactoryService,
-            ConfigurationService configurationService
-    ) {
+            RequestListenerContainerFactory requestListenerContainerFactory,
+            ConfigurationService configurationService,
+            ErrorHandlerFactory errorHandlerFactory) {
         RequestTopicNameParameters requestTopicNameParameters = RequestTopicNameParameters
                 .builder()
-                .resource("configuration")
+                .topicNamePrefixParameters(TopicNamePrefixParameters
+                        .builder()
+                        .orgIdApplicationDefault()
+                        .domainContextApplicationDefault()
+                        .build()
+                )
+                .resourceName("configuration")
                 .parameterName("configuration-id")
                 .build();
-        requestTopicService
-                .ensureTopic(requestTopicNameParameters, 0, TopicCleanupPolicyParameters.builder().build());
+        requestTopicService.createOrModifyTopic(requestTopicNameParameters, RequestTopicConfiguration
+                .builder()
+                .retentionTime(RETENTION_TIME)
+                .build()
+        );
 
-        return requestConsumerFactoryService.createRecordConsumerFactory(
+        return requestListenerContainerFactory.createRecordConsumerFactory(
                 Long.class,
                 ConfigurationDto.class,
                 (ConsumerRecord<String, Long> consumerRecord) -> ReplyProducerRecord
                         .<ConfigurationDto>builder()
                         .value(configurationService.findById(consumerRecord.value(), true).orElse(null))
-                        .build()
+                        .build(),
+                RequestListenerConfiguration
+                        .stepBuilder(Long.class)
+                        .maxPollRecordsKafkaDefault()
+                        .maxPollIntervalKafkaDefault()
+                        .build(),
+                errorHandlerFactory.createErrorHandler(
+                        ErrorHandlerConfiguration
+                                .stepBuilder()
+                                .noRetries()
+                                .skipFailedRecords()
+                                .build()
+                )
         ).createContainer(requestTopicNameParameters);
     }
 
@@ -45,18 +74,28 @@ public class ConfigurationRequestConsumerConfiguration {
     ConcurrentMessageListenerContainer<String, Long>
     mappingByConfigurationIdRequestConsumer(
             RequestTopicService requestTopicService,
-            RequestConsumerFactoryService requestConsumerFactoryService,
-            ConfigurationService configurationService
+            RequestListenerContainerFactory requestListenerContainerFactory,
+            ConfigurationService configurationService,
+            ErrorHandlerFactory errorHandlerFactory
     ) {
         RequestTopicNameParameters requestTopicNameParameters = RequestTopicNameParameters
                 .builder()
-                .resource("mapping")
+                .topicNamePrefixParameters(TopicNamePrefixParameters
+                        .builder()
+                        .orgIdApplicationDefault()
+                        .domainContextApplicationDefault()
+                        .build()
+                )
+                .resourceName("mapping")
                 .parameterName("configuration-id")
                 .build();
-        requestTopicService
-                .ensureTopic(requestTopicNameParameters, 0, TopicCleanupPolicyParameters.builder().build());
+        requestTopicService.createOrModifyTopic(requestTopicNameParameters, RequestTopicConfiguration
+                .builder()
+                .retentionTime(RETENTION_TIME)
+                .build()
+        );
 
-        return requestConsumerFactoryService.createRecordConsumerFactory(
+        return requestListenerContainerFactory.createRecordConsumerFactory(
                 Long.class,
                 ObjectMappingDto.class,
                 (ConsumerRecord<String, Long> consumerRecord) -> ReplyProducerRecord
@@ -66,7 +105,18 @@ public class ConfigurationRequestConsumerConfiguration {
                                         .findById(consumerRecord.value(), false)
                                         .map(ConfigurationDto::getMapping)
                                         .orElse(null))
-                        .build()
+                        .build(),
+                RequestListenerConfiguration
+                        .stepBuilder(Long.class)
+                        .maxPollRecordsKafkaDefault()
+                        .maxPollIntervalKafkaDefault()
+                        .build(),
+                errorHandlerFactory.createErrorHandler(
+                        ErrorHandlerConfiguration
+                                .stepBuilder()
+                                .noRetries()
+                                .skipFailedRecords().build()
+                )
         ).createContainer(requestTopicNameParameters);
     }
 
