@@ -12,6 +12,8 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 
+import static org.springframework.util.StringUtils.hasText;
+
 @Service
 @RequiredArgsConstructor
 @Slf4j
@@ -27,76 +29,71 @@ public class TokenParsingUtils {
     }
 
     private static String toTitleCase(String input) {
-        if (!StringUtils.hasText(input)) return input;
+        if (!hasText(input)) return input;
 
-        String s = input.trim();
-        StringBuilder out = new StringBuilder(s.length());
+        StringBuilder result = new StringBuilder(input.length());
         boolean startOfWord = true;
 
-        for (int i = 0; i < s.length(); i++) {
-            char c = s.charAt(i);
+        for (char c : input.trim().toCharArray()) {
             if (Character.isLetter(c)) {
-                out.append(startOfWord ? Character.toTitleCase(c) : Character.toLowerCase(c));
+                result.append(startOfWord ? Character.toTitleCase(c) : Character.toLowerCase(c));
                 startOfWord = false;
             } else {
-                out.append(c);
+                result.append(c);
                 startOfWord = Character.isWhitespace(c) || c == '-' || c == '\'' || c == '’';
             }
         }
-        return out.toString();
+        return result.toString();
     }
 
     public Optional<User> tryGetUser(Authentication authentication) {
-        if (!(authentication instanceof JwtAuthenticationToken jwt)) {
-            return Optional.empty();
-        }
+        if (!(authentication instanceof JwtAuthenticationToken jwt)) return Optional.empty();
 
-        Map<String, Object> a = jwt.getTokenAttributes();
+        Map<String, Object> attrs = jwt.getTokenAttributes();
 
-        String displayName = text(a.get("displayname"));
-        String email       = text(a.get("email"));
-        String principal   = text(a.get("principalName"));
-        String name        = firstNonBlank(
-                displayName,
-                text(a.get("name")),
-                text(a.get("preferred_username")),
-                principal
+        String name = firstNonBlank(
+                text(attrs.get("displayname")),
+                text(attrs.get("name")),
+                text(attrs.get("preferred_username")),
+                text(attrs.get("principalName"))
         );
 
-        UUID oid = parseUuid(text(a.get("objectidentifier")));
-
-        if (name == null && email == null && oid == null) {
-            return Optional.empty();
-        }
-
-        return Optional.of(
-                User.builder()
-                        .objectIdentifier(oid)
-                        .name(name)
-                        .email(email != null ? email : principal)
-                        .build()
+        String email = firstNonBlank(
+                text(attrs.get("email")),
+                text(attrs.get("principalName"))
         );
+
+        UUID oid = parseUuid(text(attrs.get("objectidentifier")));
+
+        if ((name == null && email == null) || oid == null) return Optional.empty();
+
+        return Optional.of(User.builder()
+                .objectIdentifier(oid)
+                .name(name)
+                .email(email)
+                .build());
     }
 
     private static String text(Object o) {
-        if (o == null) return null;
-        String s = StringUtils.trimWhitespace(String.valueOf(o));
-        return StringUtils.hasText(s) ? s : null;
+        return Optional.ofNullable(o)
+                .map(Object::toString)
+                .map(String::strip)
+                .filter(StringUtils::hasText)
+                .orElse(null);
     }
 
-    private static String firstNonBlank(String... vals) {
-        if (vals == null) return null;
-        for (String v : vals) {
-            if (StringUtils.hasText(v)) return v;
+    private static String firstNonBlank(String... values) {
+        if (values == null) return null;
+        for (String v : values) {
+            if (hasText(v)) return v;
         }
         return null;
     }
 
     private static UUID parseUuid(String s) {
-        if (s == null) return null;
         try {
-            return UUID.fromString(s);
-        } catch (IllegalArgumentException ignore) {
+            return s == null ? null : UUID.fromString(s);
+        } catch (IllegalArgumentException e) {
             return null;
         }
     }
