@@ -1,7 +1,7 @@
 package no.fintlabs.validation;
 
+import jakarta.validation.Payload;
 import lombok.Builder;
-import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.extern.jackson.Jacksonized;
@@ -9,7 +9,6 @@ import no.fintlabs.model.integration.Integration;
 import no.fintlabs.model.metadata.InstanceValueMetadata;
 import no.fintlabs.model.metadata.IntegrationMetadata;
 
-import javax.validation.Payload;
 import java.util.Map;
 
 @Getter
